refactor: update import paths for VerifiedBadge component across multiple files
