+++ src/app/(public)/video/[publicId]/comments/CommentItem.tsx
@@ -5,8 +5,8 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { useEffect, useRef, useState } from 'react'
 import { Heading } from '@/ui/Heading'
+import { VerifiedBadge } from '@/ui/VerifiedBadge'
 import AutoResizeTextarea from '@/ui/fields/AutoResizeTextarea'
-import { VerifiedBadge } from '@/ui/video-card/VerifiedBadge'
 import { PAGE } from '@/config/public-page.config'
 import { useAuth } from '@/hooks/useAuth'
 import { transformDate } from '@/utils/transform-date'
