+++ src/app/(public)/video/[publicId]/video-channel/VideoChannel.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { Heading } from '@/ui/Heading'
 import { SkeletonLoader } from '@/ui/SkeletonLoader'
-import { VerifiedBadge } from '@/ui/video-card/VerifiedBadge'
+import { VerifiedBadge } from '@/ui/VerifiedBadge'
 import { PAGE } from '@/config/public-page.config'
 import { transformCount } from '@/utils/transform-count'
 import type { ISingleVideoResponse } from '@/types/video.types'
