+++ src/components/ui/VerifiedBadge.tsx
